requirements.txt
pandas
langchain
openai
python-dotenv
chromadb
tiktoken
sentence-transformers
stability-sdk
